+++ apps/parsley/src/App.tsx
@@ -1,6 +1,6 @@
 import styled from "@emotion/styled";
 import { Route, BrowserRouter as Router, Routes } from "react-router-dom";
-import { AuthProvider } from "@evg-ui/lib/context/Auth";
+import { AuthProvider } from "@evg-ui/lib/context/AuthProvider";
 import LoginPage from "@evg-ui/lib/pages/LoginPage";
 import { ErrorBoundary } from "components/ErrorHandling";
 import { GlobalStyles } from "components/styles";
