+++ apps/parsley/src/components/NavBar/index.tsx
@@ -6,7 +6,7 @@ import { palette } from "@leafygreen-ui/palette";
 import Icon from "@evg-ui/lib/components/Icon";
 import { StyledLink } from "@evg-ui/lib/components/styles";
 import { size } from "@evg-ui/lib/constants/tokens";
-import { useAuthProviderContext } from "@evg-ui/lib/context/Auth";
+import { useAuthProviderContext } from "@evg-ui/lib/context/AuthProvider";
 import DetailsMenu from "components/DetailsMenu";
 import Search from "components/Search";
 import ShortcutModal from "components/ShortcutModal";
